+++ test/noahtheduke/splint/integrations/iort_test.clj
@@ -22,7 +22,8 @@
     performance/assoc-many 1
     performance/dot-equals 3
     style/first-first 1
-    style/not-some-pred 1})
+    style/not-some-pred 1
+    style/prefer-var-dispatch 7})
 
 (defdescribe iort-test
   (let [iort (delay (gl/procure "https://github.com/wardle/iort.git"
@@ -46,7 +47,7 @@
          (m/equals iort-diagnostics)
          (update-vals* @diagnostics count))))
     (it "sums correctly"
-      (expect (= 33 (count (:diagnostics @results)))))
+      (expect (= 40 (count (:diagnostics @results)))))
     (it "raises no errors"
       (expect (nil? (get diagnostics 'splint/error))))
     (it "raises no unknown errors"